[Oztechan/CCC#2122] Move :viewmodel:premium to Client
