+++ android/app/src/main/kotlin/com/oztechan/ccc/android/app/di/Koin.kt
@@ -6,7 +6,6 @@ import co.touchlab.kermit.Logger
 import com.oztechan.ccc.android.app.BuildConfig
 import com.oztechan.ccc.android.core.ad.di.androidCoreAdModule
 import com.oztechan.ccc.android.core.billing.di.androidCoreBillingModule
-import com.oztechan.ccc.android.viewmodel.premium.di.androidViewModelPremiumModule
 import com.oztechan.ccc.android.viewmodel.widget.di.androidViewModelWidgetModule
 import com.oztechan.ccc.client.configservice.ad.di.clientConfigServiceAdModule
 import com.oztechan.ccc.client.configservice.review.di.clientConfigServiceReviewModel
@@ -24,6 +23,7 @@ import com.oztechan.ccc.client.storage.calculation.di.clientStorageCalculationMo
 import com.oztechan.ccc.client.viewmodel.calculator.di.clientViewModelCalculatorModule
 import com.oztechan.ccc.client.viewmodel.currencies.di.clientViewModelCurrenciesModule
 import com.oztechan.ccc.client.viewmodel.main.di.clientViewModelMainModule
+import com.oztechan.ccc.client.viewmodel.premium.di.clientViewModelPremiumModule
 import com.oztechan.ccc.client.viewmodel.selectcurrency.di.clientViewModelSelectCurrencyModule
 import com.oztechan.ccc.client.viewmodel.settings.di.clientViewModelSettingsModule
 import com.oztechan.ccc.client.viewmodel.watchers.di.clientViewModelWatchersModule
@@ -50,7 +50,6 @@ fun initKoin(context: Context) = startKoin {
         androidCoreBillingModule,
         // ViewModel modules
         androidViewModelWidgetModule,
-        androidViewModelPremiumModule,
         // endregion
 
         // region Client modules
@@ -79,6 +78,7 @@ fun initKoin(context: Context) = startKoin {
         clientViewModelSettingsModule,
         clientViewModelSelectCurrencyModule,
         clientViewModelWatchersModule,
+        clientViewModelPremiumModule,
         // endregion
 
         // region Common modules
