+++ android/ui/mobile/src/main/kotlin/com/oztechan/ccc/android/ui/mobile/content/premium/PremiumAdapter.kt
@@ -5,8 +5,8 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import com.github.submob.basemob.adapter.BaseVBRecyclerViewAdapter
 import com.oztechan.ccc.android.ui.mobile.databinding.ItemPremiumBinding
-import com.oztechan.ccc.android.viewmodel.premium.PremiumEvent
 import com.oztechan.ccc.client.core.shared.model.PremiumType
+import com.oztechan.ccc.client.viewmodel.premium.PremiumEvent
 
 class PremiumAdapter(
     private val premiumEvent: PremiumEvent
