+++ client/viewmodel/premium/src/main/kotlin/com/oztechan/ccc/client/viewmodel/premium/di/ClientViewModelPremiumModule.kt
@@ -1,0 +1,9 @@
+package com.oztechan.ccc.client.viewmodel.premium.di
+
+import com.oztechan.ccc.client.core.viewmodel.di.viewModelDefinition
+import com.oztechan.ccc.client.viewmodel.premium.PremiumViewModel
+import org.koin.dsl.module
+
+val clientViewModelPremiumModule = module {
+    viewModelDefinition { PremiumViewModel(get()) }
+}
