+++ client/viewmodel/premium/src/test/kotlin/com/oztechan/ccc/client/viewmodel/premium/PremiumViewModelTest.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2021 Mustafa Ozhan. All rights reserved.
  */
 
-package com.oztechan.ccc.android.viewmodel.premium
+package com.oztechan.ccc.client.viewmodel.premium
 
 import co.touchlab.kermit.CommonWriter
 import co.touchlab.kermit.Logger
